+++ code/core/src/main/java/org/betonquest/betonquest/kernel/component/AsyncSaverComponent.java
@@ -1,6 +1,5 @@
 package org.betonquest.betonquest.kernel.component;
 
-import org.betonquest.betonquest.api.config.ConfigAccessor;
 import org.betonquest.betonquest.api.config.ConfigAccessorFactory;
 import org.betonquest.betonquest.api.dependency.DependencyProvider;
 import org.betonquest.betonquest.api.logger.BetonQuestLoggerFactory;
@@ -15,7 +14,6 @@ import java.util.Set;
 /**
  * The implementation of {@link AbstractCoreComponent} for {@link AsyncSaver}.
  */
-@SuppressWarnings("PMD.DoNotUseThreads")
 public class AsyncSaverComponent extends AbstractCoreComponent {
 
     /**
@@ -27,7 +25,7 @@ public class AsyncSaverComponent extends AbstractCoreComponent {
 
     @Override
     public Set<Class<?>> requires() {
-        return Set.of(Plugin.class, BetonQuestLoggerFactory.class, ConfigAccessorFactory.class, ConfigAccessor.class, Connector.class);
+        return Set.of(Plugin.class, BetonQuestLoggerFactory.class, ConfigAccessorFactory.class, Connector.class);
     }
 
     @Override
@@ -40,11 +38,9 @@ public class AsyncSaverComponent extends AbstractCoreComponent {
         final Plugin plugin = getDependency(Plugin.class);
         final BetonQuestLoggerFactory loggerFactory = getDependency(BetonQuestLoggerFactory.class);
         final ConfigAccessorFactory configAccessorFactory = getDependency(ConfigAccessorFactory.class);
-        final ConfigAccessor config = getDependency(ConfigAccessor.class);
         final Connector connector = getDependency(Connector.class);
 
-        final AsyncSaver saver = new AsyncSaver(loggerFactory.create(AsyncSaver.class, "AsyncSaver"), config.getLong("mysql.reconnect_interval"), connector);
-        new Thread(saver).start();
+        final AsyncSaver saver = new AsyncSaver(loggerFactory.create(AsyncSaver.class, "AsyncSaver"), connector);
         new Backup(loggerFactory, loggerFactory.create(Backup.class), configAccessorFactory, plugin.getDataFolder(), connector)
                 .loadDatabaseFromBackup();